move findAndClickButton to common ScriptUtils & update script syntax to JS (followups from aws#5435)
