+++ ui-tests-starter/tst-243+/software/aws/toolkits/jetbrains/uitests/ScriptUtils.kt
@@ -1,9 +1,9 @@
 // Copyright 2025 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package software.aws.toolkits.jetbrains.uitests.docTests.scripts
+package software.aws.toolkits.jetbrains.uitests
 
-// language=TS
+// language=JS
 val findAndClickButtonScript = """
 const findAndClickButton = async (
   page,
